fix: CI 빌드에러 해결 + 폴더구조
