+++ src/styles/shared.css.ts
@@ -4,5 +4,5 @@ export const sidebarWrapper = style({
   display: "flex",
   alignItems: "stretch",
 
-  minHeight: "calc(100vh - 7.2rem)",
+  minHeight: "calc(100vh - 14.4rem)",
 });
